Clarify titles under "Running Kubernetes"
Of course AWS does not run "on" CoreOS, so under "Running Kubernetes" the title must not be f. e. "AWS on CoreOS".
This is highly irritating when browsing the docs the first time.
The title should either be: "Kubernetes with AWS provider on CoreOS", or "CoreOS on AWS".
I picked the latter form.

I also changed "Vagrant or VMware" to "CoreOS on Vagrant", because the article is mostly about CoreOS, including other platforms (btw. VMWare is just a Vagrant provider in this context).
Also note that there is a little navigation bug because the article "/docs/getting-started-guides/coreos/" is linked twice: when clicked, the other link is selected too in the navigation (under certain circumstances). It may make sense to split the article in 2 files: one about Vagrant only, one about AWS and GCE.